Database-Q-n-A-agent-using-LangChain-SQL-toolkit/requirement.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.0.1
streamlit>=1.28.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
pandas>=2.0.0
numpy>=1.24.0
